+++ src/main/java/de/alpharogroup/swing/document/NumberValuesDocument.java
@@ -22,32 +22,53 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.swing.panels.tree;
+package de.alpharogroup.swing.document;
 
-import java.awt.Frame;
-
-import de.alpharogroup.layout.CloseWindow;
+import lombok.NonNull;
 
 /**
- * The test class for {@link JTreePanel}
+ * The class {@link NumberValuesDocument} can take any character that is specified in the given
+ * regular expression
  */
-public class JTreePanelTest
+public class NumberValuesDocument extends RegularExpressionDocument
 {
 
+	/** The Constant for the default regular expression. */
+	public static final String DEFAULT_REGEX = "^[0-9,;]+$";
+
+	/** The Constant serialVersionUID. */
+	private static final long serialVersionUID = 1L;
 
 	/**
-	 * The main method.
-	 *
-	 * @param args
-	 *            the arguments
+	 * Instantiates a new {@link NumberValuesDocument} object with the default regular expression
 	 */
-	public static void main(final String[] args)
+	public NumberValuesDocument()
 	{
-		final Frame frame = new Frame("JTreePanel");
-		frame.addWindowListener(new CloseWindow());
-		frame.add(new TestTreePanel());
-		frame.pack();
-		frame.setVisible(true);
+		this(NumberValuesDocument.DEFAULT_REGEX);
+	}
+
+	/**
+	 * Instantiates a new {@link NumberValuesDocument} object
+	 *
+	 * @param regex
+	 *            the regular expression
+	 */
+	public NumberValuesDocument(@NonNull String regex)
+	{
+		super(regex);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public String validate(String proposedValue) throws IllegalArgumentException
+	{
+		if (proposedValue.isEmpty())
+		{
+			return proposedValue;
+		}
+		return super.validate(proposedValue);
 	}
 
 }