+++ CHANGELOG.md
@@ -8,6 +8,8 @@ ADDED:
  
 - new JTreePanel class created for JTree
 - new RangeDocument class created for constraint the range of a document
+- new RegularExpressionDocument class created for constraint the input of a document with the given regular expression
+- new NumberValuesDocument class created for constraint the input of a document only for numbers separated with comma or semicolon
 
 CHANGED: